Version 1.0 (based on 8265d44eaa5c7bdb95ec2fe65796009c16db226f)
This is the initial public release of a framework, developed by the VMware
vCloud Director team, which allows developers to incrementally define a
database upgrade process for a Java applications.

This commit represents the work of many developers spanning several years;
while the framework is being released, its development history is not.