core: allow PersistenceContexts to be retrieved using a super type
Previously, only PersistenceContexts whose Class exactly matched the requested type were returned. With this change, all PersistenceContexts whose class is assignable to the requested type are considered as candidates.

This change allows users to, for example, extend DatabasePersistenceContext to include product-specific methods and still be able to retrieve that context via the DatabasePersistenceContext interface when that additional functionality is not required.