sql-dsl: fix column nullability bug
The call to set the nullability of a `ColumnType` was being called on an object
that could be re-used when adding custom types, which, once set, could cause
future upgrades that used that type to allow nulls (instead of the default). This
addresses that by creating a new copy of the object if it implements the
`NullAware` interface.

To avoid duplication of the copy calls, as well as to reduce existing duplication
regarding determining if a `sql` call is required and creation of the `allowing`
closure a utility class was added.