sql-dsl: validate column names for all relevant operations
Column names were only validated when adding a column to an
existing table or renaming a column. Validation will now occur
on table creation as well.

Additionally, for adding or renaming a column validation of
length and reserved keyword conflict are now both in the
syntax class (instead of one in the syntax and one in the
model).

Unit tests have been added to better cover the variety of
situations validation should occur.