dsl: add the ability to set a default value
Adds a keyword to optionally set the default value for a column - including
overriding the default for the DataTypes that have one predefined (e.g. `BOOL`).

Because in some cases (e.g. a timestamp) it is desirable for the default to
be a SQL operation instead of a set value a separate (but similar) keyword was
added to allow doing so.

This commit changes how the `BOOL` and `DATE` types are defined but does not alter
the generated SQL (i.e. using one of these types, without specifying a custom
default, will result in identical SQL to what was generated previously).