build: upgrade to java 11 and groovy 3
Update the upgrade-framework to compile using Java 11 and Groovy 3. Had to
switch to using gobuild's MakeHelper instead of MavenHelper because MavenHelper
only looks at versions of maven available in toolchain, which is not being
updated anymore.

Had to make a variety of javadoc changes to appease the newer version
maven-javadoc-plugin which seems to be more strict. I just picked the path of
least resistance in modifying them.

Also updated the project's version to 2.0.0 and replaced travis with github
actions.

Signed-off-by: Ryan Lewis <ryanlewis@vmware.com>